Update Builder.php
I search every where to fix the problem with column set as identity.  MsSQL shows this error every time that update a record using ORM: "Cannot update identity column 'ID'."
The problem is that the column ID was on the update query, so I made some changes to remove that field.  Is not a perfect way, but works for me for my project now.  check it out and I hope you can fix it as soon as possible.  Thank you.